Moved large sprite and scrolling to folders
