+++ internal/scrolling.c
